linux-gen: packet: shrink _odp_packet_input_flags_t to 32 bits
Now with fewer packet input flags, shrink _odp_packet_input_flags_t
from 8 to 4 bytes. This reduces the size of packet_parset_t by 8 bytes
since structure padding there gets reduced too. Add explicit unused
padding bytes in packet header (for now) to avoid moving other packet
header fields between cache lines.

+++ platform/linux-generic/include/odp/api/plat/packet_flag_inlines.h
@@ -22,9 +22,9 @@ extern "C" {
 
 /** @cond _ODP_HIDE_FROM_DOXYGEN_ */
 
-static inline uint64_t _odp_packet_input_flags(odp_packet_t pkt)
+static inline uint32_t _odp_packet_input_flags(odp_packet_t pkt)
 {
-	return _odp_pkt_get(pkt, uint64_t, input_flags);
+	return _odp_pkt_get(pkt, uint32_t, input_flags);
 }
 
 #ifndef _ODP_NO_INLINE
